moving geometry file validator into validators file
